Filter and update data from user
